refactor: move functions into seperate files
